+++ mithril-common/src/test_utils/mithril_fixture.rs
@@ -180,8 +180,11 @@ impl MithrilFixture {
         let genesis_avk = self.compute_avk();
         let genesis_signer = ProtocolGenesisSigner::create_deterministic_genesis_signer();
         let genesis_producer = CertificateGenesisProducer::new(Some(Arc::new(genesis_signer)));
-        let genesis_protocol_message =
-            CertificateGenesisProducer::create_genesis_protocol_message(&genesis_avk).unwrap();
+        let genesis_protocol_message = CertificateGenesisProducer::create_genesis_protocol_message(
+            &self.protocol_parameters,
+            &genesis_avk,
+        )
+        .unwrap();
         let genesis_signature = genesis_producer
             .sign_genesis_protocol_message(genesis_protocol_message)
             .unwrap();